Extract out comparator classes in SortCommand class into separate new classes
