+++ src/main/java/seedu/address/logic/commands/SortCommand.java
@@ -11,6 +11,8 @@ import java.util.Comparator;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.contact.Contact;
+import seedu.address.model.contact.comparator.DateComparator;
+import seedu.address.model.contact.comparator.NameComparator;
 
 /**
  * Sorts the list of contacts in the address book.
@@ -54,26 +56,5 @@ public class SortCommand extends Command {
         return new CommandResult(message);
     }
 
-    class NameComparator implements Comparator<Contact> {
-        @Override
-        public int compare(Contact p1, Contact p2) {
-            return p1.compareTo(p2);
-        }
-    }
 
-    class DateComparator implements Comparator<Contact> {
-        @Override
-        public int compare(Contact p1, Contact p2) {
-            Timestamp t1 = p1.getTimeAdded().getTimestamp();
-            Timestamp t2 = p2.getTimeAdded().getTimestamp();
-
-            if (t1.before(t2)) {
-                return -1;
-            } else if (t1.after(t2)) {
-                return 1;
-            } else {
-                return 0;
-            }
-        }
-    }
 }
